Adds cookie consent banner for GDPR compliance
Implements a cookie consent banner to comply with GDPR regulations.

This change introduces a new component that informs users about the use of cookies and requests their consent.
It also allows users to accept or decline the use of cookies.
The banner is displayed on all pages of the website in non-development environments.